Change github thread ID from int to long

+++ github-client/src/main/kotlin/net/plshark/github/client/notifications/NotificationsClient.kt
@@ -8,5 +8,5 @@ interface NotificationsClient {
     fun getNotifications(): Flow<Notification>
 
     /** Marks a thread as done. */
-    suspend fun markThreadDone(threadId: Int)
+    suspend fun markThreadDone(threadId: Long)
 }
